back-end/requirements.txt
fastapi
uvicorn
PyPDF2
httpx>=0.27.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
google-genai
pypdfium2
numpy
faiss-cpu
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0